Fix packer and nvim already installed check
It was not possible to install LunarVim if these programs were already
installed. The check is syntactically incorrect and will always return
non-0 status, and the installer always continued with installing packer
and nvim. This failed with an error: "fatal: destination path
'REDACTED/packer.nvim' already exists and is not an empty directory."

Correct the syntax error by checking if the file exists.