Update README.md
As I initially pointed out in #475, a lot of GH issues are probably better handled through some community. I was informed that it already existed, so this is just a update to the README pointing users to the Discord/Matrix server for help. I through in an additional two bullet points. Didn't want to be too rude, but many issues do seem to be better directed at the specific plug-in repos/communities and not here.