Added basic plugin use/load/require function
`use_maker` takes Packer's `use` and returns a function which takes
Packer's `plugin_settings` table. This function is then used to call
Packer's own `use`, then determines if the plugin is lazy loaded and if
it has a config that needs to be `require`'d. Due to execution order, we
don't call `require_plugin` or `require` immediately, but add the plugin
to a table. This table is then called at the end of `plugins.lua` and
will first "pre-load" the plugin if it's being loaded lazily (ie `opt =
true`) and it will call `require` if it was determined that this plugin
does in fact have a configuration that needs to be required.

The one major issue at the moment is that the logic used to dertermine
if `require` should be called is not very robust. It only is going to
look for directories in our `./lua` dir that are named after the
plugin, sans any extensions (ie for `nvim-treesitter/nvim-treesitter` it
looks for the `lv-nvim-treesitter` dir and for `folke/trouble.nvim` it
looks for the `lv-trouble` dir). Either more robust logic can be added
to account for a variety of potential directory names or a single
directory naming scheme would need to be adapted.

Additional work that probably needs to be done prior to a merge:

- removing `require` statements from `init.lua` for all plugins, but this
  cannot be done until the above issue is resolved
- move helper functions into a `utils` file to clean up the
  `plugins.lua` file, which in my opinion should ideally just be the
  minimum boilerplate needed to `Use` a plugin
- I'd love to make use of the Packer optionals like `config` and
  `setup`, however at the moment it seems to be an issue due to load
  order, some plugins need other plugins, not sure if this is even an
  option, but it might be worth looking into