User configurable latex lsp#2 (#861)
* Adjust tex.lua and default-config for lsp magic

With this commit table elements are added to default-config.lua, which
are referenced in tex.lua. The settings are set to default.

This enables the user to configure settings accordingly inside of
lv-config.lua.
- build arguments
- forwardsearch executable e.g. zathura
- forwardsearch arguments for chosen executable are implemented
  automatically
- handlers for virtual_text, signs, underline and update_in_insert are
  set to the default

* Added Texlab-Build and Forward as which-keys

With this commit the two commands `:TexlabBuild` and `:TexlabForward`
are added as which-keys under the category +LaTeX.

- b is for TexlabBuild
- p is for TexlanForward (Preview)

* Change which-key for +Latex

The which-key for +Latex has to be changed because `L` is already used
by +Lush.

With this commit the which-key for +Latex is changed to `t`.

* Add documentation of options

All configurable texlab option is documented in lv-config.example.lua.

Also a link to the documentation of the options from latex-lsp/texlab
is added.