feat: add possibility to override plugin configuration
It can be useful be able to override the default plugin configuration
from the user configuration. One example is to be able to change at what
event a plugin should be loaded on.

Merge all the plugin configurations by keying on the plugin name and let
the configurations down the chain take precedence by overriding the
configuration. The order of which the plugins are loaded is kept in the
order they are specified to keep it deterministic and consistent.