fix(lualine): use tabstop in spaces component when sensible
There is no guarantee that shiftwidth is always set to the same value as
tabstop.  Additionally, when shiftwidth is set to the value zero (0) the
value of tabstop is used instead.  As such, it is not useful to solely
rely on that one option.  When 'noexpandtab' is set, the shiftwidth
value becomes somewhat meaningless.

From :help 'shiftwidth':
When zero the 'ts' value will be used.