fix: adjust example of unmapping an lvim default
The example config file has a sample for unmapping a LunarVim key by
setting it to "".  When I tried that, the result was that the key became
a no-op.  If I instead set it to nil, the key returned to default vim
operation.  In particular, I was trying to unmap normal mode L and H.  I
got to nil by searching for how to delete a entry in a dictionary in
lua, but I'm not familiar with lua.