fix(autocmds): toggle format-on-save properly
`LvimToggleFormatOnSave` can disable the format-on-save feature properly, but once the feature
is disabled from the enabled state, this command cannot enable it again.
This is because `vim.api.nvim_get_autocmds` returns an empty list after disabling the feature.

This commit checks the size of the returned autocmd list when `M.toolge_format_on_save` is
called and enables the feature if the size is 0.