fix(core.comment): fix default extra mappings (#2768)
Comment default config was set with extra = false with a comment
suggesting this was for `g>`, `g<`, etc. mappings. However the `extra`
argument is for the `gco`, `gcO`, and `gcA` mappings which are enabled
by default and, in my opinion are very useful. The extended mappings are
still disabled by default.