fix: lunarvim_initialization flowchart background
When viewing the Docs > For Developers > Under the Hood page on the
website, the flowchart image is rendered with a white background.
It doesn't look very nice on the dark theme.
This patch is a remake of the same flowchart using mermaid.live tools
with a fully transparent background and colored lines that are visible
on both dark and light themes